// convex/_generated/api.d.ts
/* eslint-disable */
/**
 * Generated `api` utility.
 *
 * THIS CODE IS AUTOMATICALLY GENERATED.
 *
 * Generated by convex@1.14.4.
 * To regenerate, run `npx convex dev`.
 * @module
 */

import type {
  ApiFromModules,
  FilterApi,
  FunctionReference,
} from "convex/server";
import type * as actions from "../actions.js";
import type * as appNamespaces from "../appNamespaces.js";
import type * as appTemplates from "../appTemplates.js";
import type * as appVersions from "../appVersions.js";
import type * as apps from "../apps.js";
import type * as clients from "../clients.js";
import type * as connections from "../connections.js";
import type * as functions from "../functions.js";
import type * as helpers from "../helpers.js";
import type * as http from "../http.js";
import type * as sessions from "../sessions.js";
import type * as stripe from "../stripe.js";
import type * as users from "../users.js";

/**
 * A utility for referencing Convex functions in your app's API.
 *
 * Usage:
 * ```js
 * const myFunctionReference = api.myModule.myFunction;
 * ```
 */
declare const fullApi: ApiFromModules<{
  actions: typeof actions;
  appNamespaces: typeof appNamespaces;
  appTemplates: typeof appTemplates;
  appVersions: typeof appVersions;
  apps: typeof apps;
  clients: typeof clients;
  connections: typeof connections;
  functions: typeof functions;
  helpers: typeof helpers;
  http: typeof http;
  sessions: typeof sessions;
  stripe: typeof stripe;
  users: typeof users;
}>;
export declare const api: FilterApi<
  typeof fullApi,
  FunctionReference<any, "public">
>;
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;